added AbstractBaseBarSeries for those wishing to extend BasBar and continue to use BaseBarSeries
