+++ ta4j-core/src/main/java/org/ta4j/core/BaseBarSeries.java
@@ -23,214 +23,47 @@
  */
 package org.ta4j.core;
 
-import static org.ta4j.core.num.NaN.NaN;
+import org.ta4j.core.num.Num;
 
 import java.math.BigDecimal;
 import java.time.Duration;
 import java.time.ZonedDateTime;
-import java.util.ArrayList;
 import java.util.List;
-import java.util.Objects;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.ta4j.core.num.DecimalNum;
-import org.ta4j.core.num.Num;
 
 /**
  * Base implementation of a {@link BarSeries}.
  * </p>
  */
-public class BaseBarSeries implements BarSeries<BaseBar> {
+public class BaseBarSeries extends BaseBarSeriesAbstract<BaseBar> {
 
     private static final long serialVersionUID = -1878027009398790126L;
 
-    /**
-     * The logger
-     */
-    private final transient Logger log = LoggerFactory.getLogger(getClass());
-
-    /**
-     * Name for unnamed series
-     */
-    private static final String UNNAMED_SERIES_NAME = "unnamed_series";
-
-    /**
-     * Any instance of Num to determine its Num type.
-     */
-    protected final transient Num num;
-
-    /**
-     * Name of the series
-     */
-    private final String name;
-    /**
-     * List of bars
-     */
-    private final List<BaseBar> bars;
-    /**
-     * Begin index of the bar series
-     */
-    private int seriesBeginIndex;
-    /**
-     * End index of the bar series
-     */
-    private int seriesEndIndex;
-    /**
-     * Maximum number of bars for the bar series
-     */
-    private int maximumBarCount = Integer.MAX_VALUE;
-    /**
-     * Number of removed bars
-     */
-    private int removedBarsCount = 0;
-    /**
-     * True if the current series is constrained (i.e. its indexes cannot change),
-     * false otherwise
-     */
-    private final boolean constrained;
-
-    /**
-     * Constructor of an unnamed series.
-     */
     public BaseBarSeries() {
-        this(UNNAMED_SERIES_NAME);
+        super();
     }
 
-    /**
-     * Constructor.
-     *
-     * @param name the name of the series
-     */
     public BaseBarSeries(String name) {
-        this(name, new ArrayList<>());
+        super(name);
     }
 
-    /**
-     * Constructor of an unnamed series.
-     *
-     * @param bars the list of bars of the series
-     */
     public BaseBarSeries(List<BaseBar> bars) {
-        this(UNNAMED_SERIES_NAME, bars);
+        super(bars);
     }
 
-    /**
-     * Constructor.
-     *
-     * @param name the name of the series
-     * @param bars the list of bars of the series
-     */
     public BaseBarSeries(String name, List<BaseBar> bars) {
-        this(name, bars, 0, bars.size() - 1, false);
+        super(name, bars);
     }
 
-    /**
-     * Constructor.
-     * 
-     * @param name the name of the series
-     * @param num  any instance of Num to determine its Num function; with this, we
-     *             can convert a {@link Number} to a {@link Num Num implementation}
-     */
     public BaseBarSeries(String name, Num num) {
-        this(name, new ArrayList<>(), num);
+        super(name, num);
     }
 
-    /**
-     * Constructor.
-     *
-     * @param name the name of the series
-     * @param bars the list of bars of the series
-     * @param num  any instance of Num to determine its Num function; with this, we
-     *             can convert a {@link Number} to a {@link Num Num implementation}
-     */
     public BaseBarSeries(String name, List<BaseBar> bars, Num num) {
-        this(name, bars, 0, bars.size() - 1, false, num);
+        super(name, bars, num);
     }
 
-    /**
-     * Constructor.
-     * <p/>
-     * Creates a BaseBarSeries with default {@link DecimalNum} as type for the data
-     * and all operations on it
-     *
-     * @param name             the name of the series
-     * @param bars             the list of bars of the series
-     * @param seriesBeginIndex the begin index (inclusive) of the bar series
-     * @param seriesEndIndex   the end index (inclusive) of the bar series
-     * @param constrained      true to constrain the bar series (i.e. indexes cannot
-     *                         change), false otherwise
-     */
-    private BaseBarSeries(String name, List<BaseBar> bars, int seriesBeginIndex, int seriesEndIndex, boolean constrained) {
-        this(name, bars, seriesBeginIndex, seriesEndIndex, constrained, DecimalNum.ZERO);
-    }
-
-    /**
-     * Constructor.
-     *
-     * @param name             the name of the series
-     * @param bars             the list of bars of the series
-     * @param seriesBeginIndex the begin index (inclusive) of the bar series
-     * @param seriesEndIndex   the end index (inclusive) of the bar series
-     * @param constrained      true to constrain the bar series (i.e. indexes cannot
-     *                         change), false otherwise
-     * @param num              any instance of Num to determine its Num function;
-     *                         with this, we can convert a {@link Number} to a
-     *                         {@link Num Num implementation}
-     */
-    BaseBarSeries(String name, List<BaseBar> bars, int seriesBeginIndex, int seriesEndIndex, boolean constrained, Num num) {
-        this.name = name;
-
-        this.bars = bars;
-        if (bars.isEmpty()) {
-            // Bar list empty
-            this.seriesBeginIndex = -1;
-            this.seriesEndIndex = -1;
-            this.constrained = false;
-            this.num = num;
-            return;
-        }
-        // Bar list not empty: take Function of first bar
-        this.num = bars.get(0).getClosePrice();
-        // Bar list not empty: checking num types
-        if (!checkBars(bars)) {
-            throw new IllegalArgumentException(String.format(
-                    "Num implementation of bars: %s" + " does not match to Num implementation of bar series: %s",
-                    bars.get(0).getClosePrice().getClass(), num.function()));
-        }
-        // Bar list not empty: checking indexes
-        if (seriesEndIndex < seriesBeginIndex - 1) {
-            throw new IllegalArgumentException("End index must be >= to begin index - 1");
-        }
-        if (seriesEndIndex >= bars.size()) {
-            throw new IllegalArgumentException("End index must be < to the bar list size");
-        }
-        this.seriesBeginIndex = seriesBeginIndex;
-        this.seriesEndIndex = seriesEndIndex;
-        this.constrained = constrained;
-    }
-
-    /**
-     * Cuts a list of bars into a new list of bars that is a subset of it
-     *
-     * @param bars       the list of {@link Bar bars}
-     * @param startIndex start index of the subset
-     * @param endIndex   end index of the subset
-     * @return a new list of bars with tick from startIndex (inclusive) to endIndex
-     *         (exclusive)
-     */
-    private static List<BaseBar> cut(List<BaseBar> bars, final int startIndex, final int endIndex) {
-        return new ArrayList<>(bars.subList(startIndex, endIndex));
-    }
-
-    /**
-     * @param series a bar series
-     * @param index  an out of bounds bar index
-     * @return a message for an OutOfBoundsException
-     */
-    private static String buildOutOfBoundsMessage(BaseBarSeries series, int index) {
-        return String.format("Size of series: %s bars, %s bars removed, index = %s", series.bars.size(),
-                series.removedBarsCount, index);
+    public BaseBarSeries(String name, List<BaseBar> bars, int seriesBeginIndex, int seriesEndIndex, boolean constrained, Num num) {
+        super(name, bars, seriesBeginIndex, seriesEndIndex, constrained, num);
     }
 
     /**
@@ -266,158 +99,6 @@ public class BaseBarSeries implements BarSeries<BaseBar> {
         }
         return new BaseBarSeries(name, num);
 
-    }
-
-    @Override
-    public Num num() {
-        return num;
-    }
-
-    /**
-     * Checks if all {@link Bar bars} of a list fits to the {@link Num NumFunction}
-     * used by this bar series.
-     *
-     * @param bars a List of Bar objects.
-     * @return false if a Num implementation of at least one Bar does not fit.
-     */
-    private boolean checkBars(List<BaseBar> bars) {
-        for (Bar bar : bars) {
-            if (!checkBar(bar)) {
-                return false;
-            }
-        }
-        return true;
-    }
-
-    /**
-     * Checks if the {@link Num} implementation of a {@link Bar} fits to the
-     * NumFunction used by bar series.
-     *
-     * @param bar a Bar object.
-     * @return false if another Num implementation is used than by this bar series.
-     * @see Num
-     * @see Bar
-     * @see #addBar(Duration, ZonedDateTime)
-     */
-    private boolean checkBar(Bar bar) {
-        if (bar.getClosePrice() == null) {
-            return true; // bar has not been initialized with data (uses deprecated constructor)
-        }
-        // all other constructors initialize at least the close price, check if Num
-        // implementation fits to numFunction
-        Class<? extends Num> f = one().getClass();
-        return f == bar.getClosePrice().getClass() || bar.getClosePrice().equals(NaN);
-    }
-
-    @Override
-    public String getName() {
-        return name;
-    }
-
-    @Override
-    public BaseBar getBar(int i) {
-        int innerIndex = i - removedBarsCount;
-        if (innerIndex < 0) {
-            if (i < 0) {
-                // Cannot return the i-th bar if i < 0
-                throw new IndexOutOfBoundsException(buildOutOfBoundsMessage(this, i));
-            }
-            if (log.isTraceEnabled()) {
-                log.trace("Bar series `{}` ({} bars): bar {} already removed, use {}-th instead", name, bars.size(), i,
-                        removedBarsCount);
-            }
-            if (bars.isEmpty()) {
-                throw new IndexOutOfBoundsException(buildOutOfBoundsMessage(this, removedBarsCount));
-            }
-            innerIndex = 0;
-        } else if (innerIndex >= bars.size()) {
-            // Cannot return the n-th bar if n >= bars.size()
-            throw new IndexOutOfBoundsException(buildOutOfBoundsMessage(this, i));
-        }
-        return bars.get(innerIndex);
-    }
-
-    @Override
-    public int getBarCount() {
-        if (seriesEndIndex < 0) {
-            return 0;
-        }
-        final int startIndex = Math.max(removedBarsCount, seriesBeginIndex);
-        return seriesEndIndex - startIndex + 1;
-    }
-
-    @Override
-    public List<BaseBar> getBarData() {
-        return bars;
-    }
-
-    @Override
-    public int getBeginIndex() {
-        return seriesBeginIndex;
-    }
-
-    @Override
-    public int getEndIndex() {
-        return seriesEndIndex;
-    }
-
-    @Override
-    public int getMaximumBarCount() {
-        return maximumBarCount;
-    }
-
-    @Override
-    public void setMaximumBarCount(int maximumBarCount) {
-        if (constrained) {
-            throw new IllegalStateException("Cannot set a maximum bar count on a constrained bar series");
-        }
-        if (maximumBarCount <= 0) {
-            throw new IllegalArgumentException("Maximum bar count must be strictly positive");
-        }
-        this.maximumBarCount = maximumBarCount;
-        removeExceedingBars();
-    }
-
-    @Override
-    public int getRemovedBarsCount() {
-        return removedBarsCount;
-    }
-
-    /**
-     * @param bar the <code>Bar</code> to be added
-     * @apiNote to add bar data directly use #addBar(Duration, ZonedDateTime, Num,
-     *          Num, Num, Num, Num)
-     * @throws NullPointerException if bar is null
-     */
-    @Override
-    public void addBar(BaseBar bar, boolean replace) {
-        Objects.requireNonNull(bar, "bar must not be null");
-        if (!checkBar(bar)) {
-            throw new IllegalArgumentException(
-                    String.format("Cannot add Bar with data type: %s to series with data" + "type: %s",
-                            bar.getClosePrice().getClass(), one().getClass()));
-        }
-        if (!bars.isEmpty()) {
-            if (replace) {
-                bars.set(bars.size() - 1, bar);
-                return;
-            }
-            final int lastBarIndex = bars.size() - 1;
-            ZonedDateTime seriesEndTime = bars.get(lastBarIndex).getEndTime();
-            if (!bar.getEndTime().isAfter(seriesEndTime)) {
-                throw new IllegalArgumentException(
-                        String.format("Cannot add a bar with end time:%s that is <= to series end time: %s",
-                                bar.getEndTime(), seriesEndTime));
-            }
-        }
-
-        bars.add(bar);
-        if (seriesBeginIndex == -1) {
-            // Begin index set to 0 only if it wasn't initialized
-            seriesBeginIndex = 0;
-        }
-        seriesEndIndex++;
-        removeExceedingBars();
     }
 
     @Override
@@ -458,34 +139,6 @@ public class BaseBarSeries implements BarSeries<BaseBar> {
     @Override
     public void addTrade(String price, String amount) {
         addTrade(numOf(new BigDecimal(price)), numOf(new BigDecimal(amount)));
-    }
-
-    @Override
-    public void addTrade(Num tradeVolume, Num tradePrice) {
-        getLastBar().addTrade(tradeVolume, tradePrice);
-    }
-
-    @Override
-    public void addPrice(Num price) {
-        getLastBar().addPrice(price);
-    }
-
-    /**
-     * Removes the N first bars which exceed the maximum bar count.
-     */
-    private void removeExceedingBars() {
-        int barCount = bars.size();
-        if (barCount > maximumBarCount) {
-            // Removing old bars
-            int nbBarsToRemove = barCount - maximumBarCount;
-            if (nbBarsToRemove == 1) {
-                bars.remove(0);
-            } else {
-                bars.subList(0, nbBarsToRemove).clear();
-            }
-            // Updating removed bars count
-            removedBarsCount += nbBarsToRemove;
-        }
     }
 
 }
